Put connector inside the FioRequestConfirmationScene2.js file Remove FioRequestConfirmationConnector.js
