+++ src/components/Main.ui.js
@@ -29,6 +29,7 @@ import { CreateWalletSelectFiatScene } from '../components/scenes/CreateWalletSe
 import { CryptoExchangeScene } from '../components/scenes/CryptoExchangeScene.js'
 import { CurrencySettingsScene } from '../components/scenes/CurrencySettingsScene.js'
 import { DefaultFiatSettingScene } from '../components/scenes/DefaultFiatSettingScene.js'
+import { FioRequestConfirmationScene } from '../components/scenes/FioRequestConfirmationScene2.js'
 import { PromotionSettingsScene } from '../components/scenes/PromotionSettingsScene.js'
 import { SwapSettingsScene } from '../components/scenes/SwapSettingsScene.js'
 import { WalletListScene } from '../components/scenes/WalletListScene.js'
@@ -52,7 +53,6 @@ import { FioConnectWalletConfirmConnector } from '../connectors/scenes/FioConnec
 import { FioConnectWalletConnector } from '../connectors/scenes/FioConnectWalletConnector'
 import { FioDisconnectWalletsConnector } from '../connectors/scenes/FioDisconnectWalletsConnector'
 import { FioPendingRequestConnector } from '../connectors/scenes/FioPendingRequestConnector'
-import { FioRequestConfirmationConnector } from '../connectors/scenes/FioRequestConfirmationConnector.js'
 import { FioRequestListConnector } from '../connectors/scenes/FioRequestListConnector'
 import { FioSentRequestConnector } from '../connectors/scenes/FioSentRequestConnector'
 import ManageTokens from '../connectors/scenes/ManageTokensConnector.js'
@@ -347,7 +347,7 @@ export class MainComponent extends Component<Props> {
               <Scene
                 key={Constants.FIO_REQUEST_CONFIRMATION}
                 navTransparent
-                component={ifLoggedIn(FioRequestConfirmationConnector)}
+                component={ifLoggedIn(FioRequestConfirmationScene)}
                 renderTitle={this.renderTitle(s.strings.fio_confirm_request_header)}
                 renderLeftButton={this.renderBackButton(BACK)}
                 renderRightButton={this.renderMenuButton()}
